Applying Jon's patch to fix warnings
(bzr r168)

+++ src/help.c
@@ -47,7 +47,9 @@ sp_help_about (void)
 
 	w = gtk_window_new (GTK_WINDOW_TOPLEVEL);
 	gtk_window_set_title (GTK_WINDOW (w), _("About inkscape"));
-	gtk_window_set_default_size (GTK_WINDOW (w), sp_document_width (doc), sp_document_height (doc));
+	gtk_window_set_default_size (GTK_WINDOW (w), 
+				     (gint)sp_document_width (doc), 
+				     sp_document_height (doc));
 #if 1
 	gtk_window_set_policy (GTK_WINDOW (w), TRUE, TRUE, FALSE);
 #endif
